Feat: add router for views; Refactor: code refactor for server and router
